% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomial.R
\name{BetaBinomial}
\alias{BetaBinomial}
\title{Beta Binomial family}
\usage{
BetaBinomial(count.link = "logit")
}
\arguments{
\item{count.link}{link function for the count component}
}
\description{
This is part of the new implementation.
}
