% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{Binomial}
\alias{Binomial}
\title{Binomial family}
\usage{
Binomial(count.link = "logit")
}
\arguments{
\item{count.link}{link function for the count component}
}
\description{
This is part of the new implementation.
}
