% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.control.R
\name{fitdist.control}
\alias{fitdist.control}
\title{Extra arguments for fitdist}
\usage{
fitdist.control(
  method = "BFGS",
  maxit = 10000,
  trace = FALSE,
  start = NULL,
  ...
)
}
\arguments{
\item{method}{Optimization method}

\item{maxit}{Maximum number of iterations}

\item{trace}{Verbose T/F}

\item{start}{Initialization parameters}
}
\value{
the extra arguments
}
\description{
This is part of the new implementation.
}
