% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{fitdist}
\alias{fitdist}
\title{Fit an arbitrary distribution to data}
\usage{
fitdist(
  X,
  Y,
  Z = NULL,
  offsetx = NULL,
  offsetz = NULL,
  weights = NULL,
  family,
  control = fitdist.control(...),
  ...
)
}
\arguments{
\item{X}{model matrix for the count component}

\item{Y}{response vector}

\item{Z}{model matrix for the zero component}

\item{offsetx}{offset for the count component}

\item{offsetz}{offset for the zero component}

\item{weights}{weights for each observation}

\item{family}{specify the distribution to fit}

\item{control}{extra arguments for the fitting process}
}
\value{
list of likelihood and permutation based p-values
}
\description{
This is part of the new implementation.
}
