% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.fit.fixed.R
\name{glm.fit.fixed}
\alias{glm.fit.fixed}
\title{Extend the base GLM object with the likelihood attribute}
\usage{
glm.fit.fixed(x, y, ...)
}
\arguments{
\item{x}{Model matrix}

\item{y}{Response vector}
}
\value{
the fitted model
}
\description{
This is part of the new implementation.
}
