% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_offset_2.R
\name{model_offset_2}
\alias{model_offset_2}
\title{Return an offset from model matrix}
\usage{
model_offset_2(x, terms = NULL, offset = TRUE)
}
\arguments{
\item{x}{model matrix}

\item{terms}{terms}

\item{offset}{offset}
}
\value{
offset term
}
\description{
This is part of the new implementation.
}
