% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegativeBinomial.R
\name{NegativeBinomial}
\alias{NegativeBinomial}
\title{Negative Binomial family}
\usage{
NegativeBinomial(count.link = "log")
}
\arguments{
\item{count.link}{link function for the count component}
}
\description{
This is part of the new implementation.
}
