% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\name{Poisson}
\alias{Poisson}
\title{Poisson family}
\usage{
Poisson(count.link = "log")
}
\arguments{
\item{count.link}{link function for the count component}
}
\description{
This is part of the new implementation.
}
