% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prr.test.glm.R
\name{prr.test.glm}
\alias{prr.test.glm}
\title{New Permutation of Regression Residuals (PRR) test}
\usage{
prr.test.glm(
  formula,
  data,
  family = gaussian,
  subset,
  weights,
  na.action,
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  control = glm.control(...),
  offset,
  contrasts = NULL,
  test.var = NULL,
  n.rep = 100,
  seed = 12345,
  model = TRUE,
  method = "glm.fit",
  x = FALSE,
  y = TRUE,
  ...
)
}
\arguments{
\item{formula}{Formula that specifies the model}

\item{data}{Data frame with the data}

\item{family}{Specify the distribution to fit}

\item{subset}{Take a given subset of the observations}

\item{weights}{Weights to use for the observations}

\item{na.action}{what to do with NA values}

\item{offset}{offset to use for each observation}

\item{contrasts}{contrasts to use for the model matrix}

\item{test.var}{variable in the formula to permute}

\item{n.rep}{number of permutation repetitions}

\item{method}{function that fits the model}
}
\value{
list of likelihood and permutation based p-values
}
\description{
This is part of the new implementation.
}
