% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prr.test.ll.R
\name{prr.test.ll}
\alias{prr.test.ll}
\title{New Permutation of Regression Residuals (PRR) test}
\usage{
prr.test.ll(
  formula,
  data,
  family = Poisson,
  subset,
  weights,
  na.action,
  offset,
  test.var = NULL,
  test = "both",
  n.rep = 100,
  control = fitdist.control(...),
  xyz = F,
  ...
)
}
\arguments{
\item{formula}{Formula that specifies the model}

\item{data}{Data frame with the data}

\item{family}{Specify the distribution to fit}

\item{subset}{Take a given subset of the observations}

\item{weights}{Weights to use for the observations}

\item{na.action}{what to do with NA values}

\item{offset}{offset to use for each observation}

\item{test.var}{variable in the formula to permute}

\item{test}{specify whether to test "both"/"count"/"zero"}

\item{n.rep}{number of permutation repetitions}

\item{control}{fitdist.control object to adjust the optimization}

\item{xyz}{return implicit model matrices T/F}
}
\value{
list of likelihood and permutation based p-values
}
\description{
This is part of the new implementation.
}
