% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prr.test.R
\name{prr.test}
\alias{prr.test}
\title{Original Permutation of Regression Residuals (PRR) test}
\usage{
prr.test(
  formula,
  var,
  family = gaussian,
  data,
  nrep = 1000,
  seed = 12345,
  Silent = TRUE,
  weights,
  subset,
  na.action,
  start = NULL,
  etastart,
  mustart,
  offset,
  control = glm.control(...),
  model = TRUE,
  method = "glm.fit",
  x = FALSE,
  y = TRUE,
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{Formula that specifies the model}

\item{var}{variable in the formula to permute}

\item{family}{the GLM family object}

\item{data}{Data frame with the data}

\item{nrep}{number of permutation repetitions}

\item{seed}{random seed}

\item{Silent}{inverse verbose T/F}

\item{weights}{Weights to use for the observations}

\item{subset}{Take a given subset of the observations}

\item{na.action}{what to do with NA values}

\item{start}{GLM initialization parameters}

\item{etastart}{GLM initialization parameters}

\item{mustart}{GLM initialization parameters}

\item{offset}{offset to use for each observation}

\item{control}{extra fit options for GLM}

\item{model}{whether to return the model}

\item{method}{function that fits the model}

\item{x}{whether to return the model matrix}

\item{y}{whether to return the response vector}

\item{contrasts}{contrasts to use for the model matrix}
}
\value{
list of likelihood and permutation based p-values
}
\description{
This is part of the implementation in the original glmperm package (GPL-3)
}
