% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_2.R
\name{start_2}
\alias{start_2}
\title{Starting values for (ZI)Binomial and (ZI)BetaBinomial}
\usage{
start_2(
  X,
  Y,
  Z = NULL,
  offsetx = 0,
  offsetz = 0,
  weights = 1,
  zero.inflated = FALSE,
  over.dispersed = FALSE
)
}
\arguments{
\item{X}{model matrix for the count component}

\item{Y}{response vector}

\item{Z}{model matrix for the zero component}

\item{offsetx}{offset for the count component}

\item{offsetz}{offset for the zero component}

\item{weights}{weights for each observation}

\item{zero.inflated}{calculate starting values for the zero component?}

\item{over.dispersed}{calculate starting values for the dispersion parameter?}
}
\value{
list of starting values for each component, and possible over-dispersion
}
\description{
This is part of the new implementation.
}
