% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.prr.test.glm.R
\name{summary.prr.test.glm}
\alias{summary.prr.test.glm}
\title{Print PRR test statistics (GLM)}
\usage{
\method{summary}{prr.test.glm}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{the fitted PRR test object}

\item{digits}{number of digits to print}
}
\description{
This is part of the new implementation
}
