% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.prr.test.R
\name{summary.prr.test}
\alias{summary.prr.test}
\title{Print PRR test statistics}
\usage{
\method{summary}{prr.test}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{the fitted PRR test object}

\item{digits}{number of digits to print}
}
\value{
a vector with (deviance, dispersion)
}
\description{
This is part of the implementation in the original glmperm package (GPL-3)
}
