% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIBetaBinomial.R
\name{ZIBetaBinomial}
\alias{ZIBetaBinomial}
\title{Zero-Inflated Beta Binomial family}
\usage{
ZIBetaBinomial(count.link = "logit", zero.link = "logit")
}
\arguments{
\item{count.link}{link function for the count component}

\item{zero.link}{link function for the zero component}
}
\description{
This is part of the new implementation.
}
