% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIBinomial.R
\name{ZIBinomial}
\alias{ZIBinomial}
\title{Zero-Inflated Binomial family}
\usage{
ZIBinomial(count.link = "logit", zero.link = "logit")
}
\arguments{
\item{count.link}{link function for the count component}

\item{zero.link}{link function for the zero component}
}
\description{
This is part of the new implementation.
}
