% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINegativeBinomial.R
\name{ZINegativeBinomial}
\alias{ZINegativeBinomial}
\title{Zero-Inflated Negative Binomial family}
\usage{
ZINegativeBinomial(count.link = "log", zero.link = "logit")
}
\arguments{
\item{count.link}{link function for the count component}

\item{zero.link}{link function for the zero component}
}
\description{
This is part of the new implementation.
}
