% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPoisson.R
\name{ZIPoisson}
\alias{ZIPoisson}
\title{Zero-Inflated Poisson family}
\usage{
ZIPoisson(count.link = "log", zero.link = "logit")
}
\arguments{
\item{count.link}{link function for the count component}

\item{zero.link}{link function for the zero component}
}
\description{
This is part of the new implementation.
}
